Add Product service response code ok.
